+++ src/lib/pages/pool/components/PoolAssetDetail.tsx
@@ -1,4 +1,5 @@
 import { Flex, Heading, Badge, Text, Box } from "@chakra-ui/react";
+import type { Coin } from "@cosmjs/stargate";
 import type { Big } from "big.js";
 import big from "big.js";
 import Link from "next/link";
@@ -7,14 +8,14 @@ import { CustomIcon } from "lib/components/icon";
 import { Loading } from "lib/components/Loading";
 import { useAssetInfos } from "lib/services/assetService";
 import type { USD } from "lib/types";
-import type { PoolDetail, PoolLiquidity } from "lib/types/pool";
+import type { PoolDetail } from "lib/types/pool";
 import { calAssetValueWithPrecision, formatPrice } from "lib/utils";
 
 import { PoolAssetsTable } from "./table/PoolAssetsTable";
 
 interface PoolAssetDetailProps {
-  assets: PoolLiquidity[];
-  pool_type: PoolDetail["pool_type"];
+  assets: Coin[];
+  pool_type: PoolDetail["type"];
   // weight?: Option<PoolWeight[] | null>;
   // scaling_factors?: Option<string[] | null>;
 }