+++ src/lib/query/tx.ts
@@ -46,3 +46,42 @@ export const getTxsCountByAddress = graphql(`
     }
   }
 `);
+
+export const getTxsByPoolIdPagination = graphql(`
+  query getTxsByPoolIdPagination(
+    $expression: pool_transactions_bool_exp
+    $offset: Int!
+    $pageSize: Int!
+  ) {
+    pool_transactions(
+      where: $expression
+      order_by: { block_height: desc }
+      offset: $offset
+      limit: $pageSize
+    ) {
+      block {
+        height
+        timestamp
+      }
+      transaction {
+        account {
+          address
+        }
+        hash
+        success
+        messages
+        is_ibc
+      }
+    }
+  }
+`);
+
+export const getTxsCountByPoolId = graphql(`
+  query getTxsCountByPoolId($expression: pool_transactions_bool_exp) {
+    pool_transactions_aggregate(where: $expression) {
+      aggregate {
+        count
+      }
+    }
+  }
+`);