+++ src/lib/services/txService.ts
@@ -7,18 +7,31 @@ import { useQuery } from "@tanstack/react-query";
 import { useCallback } from "react";
 
 import { useCelatoneApp, useChainId } from "lib/app-provider";
-import { getTxsByAddressPagination, getTxsCountByAddress } from "lib/query";
-import type { Addr, Option, Transaction, TxFilters } from "lib/types";
+import {
+  getTxsByAddressPagination,
+  getTxsByPoolIdPagination,
+  getTxsCountByAddress,
+  getTxsCountByPoolId,
+} from "lib/query";
+import type {
+  Addr,
+  Option,
+  PoolTxFilter,
+  Transaction,
+  TxFilters,
+} from "lib/types";
+import { MsgFurtherAction, ActionMsgType } from "lib/types";
 import {
   getActionMsgType,
   getMsgFurtherAction,
   isTxHash,
+  parseDate,
   parseDateOpt,
   parseTxHash,
   snakeToCamel,
 } from "lib/utils";
 
-import { useTxExpression } from "./expression";
+import { usePoolTxExpression, useTxExpression } from "./expression";
 import type { TxResponse } from "./tx";
 import { queryTxData } from "./tx";
 
@@ -158,6 +171,82 @@ export const useTxsCountByAddress = (
     queryFn,
     {
       enabled: !!address,
+    }
+  );
+};
+
+export const useTxsByPoolIdPagination = (
+  poolId: Option<number>,
+  type: PoolTxFilter,
+  offset: number,
+  pageSize: number
+): UseQueryResult<Transaction[]> => {
+  const { indexerGraphClient } = useCelatoneApp();
+  const expression = usePoolTxExpression(poolId, type);
+
+  const queryFn = useCallback(async () => {
+    return indexerGraphClient
+      .request(getTxsByPoolIdPagination, {
+        expression,
+        offset,
+        pageSize,
+      })
+      .then(({ pool_transactions }) =>
+        pool_transactions.map((transaction) => ({
+          hash: parseTxHash(transaction.transaction.hash),
+          messages: snakeToCamel(transaction.transaction.messages),
+          sender: transaction.transaction.account.address as Addr,
+          isSigner: true,
+          height: transaction.block.height,
+          created: parseDate(transaction.block.timestamp),
+          success: transaction.transaction.success,
+          actionMsgType: ActionMsgType.OTHER_ACTION_MSG,
+          furtherAction: MsgFurtherAction.NONE,
+          isIbc: transaction.transaction.is_ibc,
+          isInstantiate: false,
+        }))
+      );
+  }, [expression, indexerGraphClient, offset, pageSize]);
+
+  return useQuery(
+    [
+      "transactions_by_pool_id",
+      poolId,
+      type,
+      offset,
+      pageSize,
+      indexerGraphClient,
+    ],
+    queryFn,
+    {
+      enabled: !!poolId,
+    }
+  );
+};
+
+export const useTxsCountByPoolId = (
+  poolId: Option<number>,
+  type: PoolTxFilter
+): UseQueryResult<Option<number>> => {
+  const { indexerGraphClient } = useCelatoneApp();
+  const expression = usePoolTxExpression(poolId, type);
+
+  const queryFn = useCallback(async () => {
+    return indexerGraphClient
+      .request(getTxsCountByPoolId, {
+        expression,
+      })
+      .then(
+        ({ pool_transactions_aggregate }) =>
+          pool_transactions_aggregate.aggregate?.count
+      );
+  }, [expression, indexerGraphClient]);
+
+  return useQuery(
+    ["transactions_count_by_pool_id", poolId, type, indexerGraphClient],
+    queryFn,
+    {
+      enabled: !!poolId,
     }
   );
 };