+++ src/lib/types/pool.ts
@@ -1,42 +1,33 @@
-import type { Option } from "lib/types";
+import type { Coin } from "@cosmjs/stargate";
+import type { Big } from "big.js";
 
-export interface PoolLiquidity {
-  denom: string;
-  amount: string;
-}
+import type { Addr, ContractAddr, Option, TokenWithValue } from "lib/types";
 
-export interface PoolWeight {
+export interface PoolWeight<T extends string | Big> {
   denom: string;
-  weight: string;
-}
-export interface TotalShares {
-  denom: string;
-  amount: string;
+  weight: T;
 }
 
-export interface SmoothWeightChangeParams {
-  start_time: string;
-  duration: string;
-}
-export interface PoolAccount {
-  address: string;
+export interface Pool<TLiquidity extends Coin | TokenWithValue = Coin> {
   id: number;
+  type: string;
+  isSuperfluid: boolean;
+  poolLiquidity: TLiquidity[];
 }
-export interface PoolDetail {
-  pool_id: number;
-  create_tx_id: number;
-  account: PoolAccount;
-  pool_liquidity: PoolLiquidity[];
-  pool_type: string;
-  is_superfluid: boolean;
-  is_supported: boolean;
-  swap_fee: string;
-  exit_fee: string;
-  future_pool_governor: Option<string | null>;
-  smooth_weight_change_params: Option<object | null>;
-  scaling_factors: Option<string[] | null>;
-  scaling_factor_controller: Option<string | null>;
-  weight: Option<PoolWeight[] | null>;
-  address: string;
-  total_shares: TotalShares;
+
+export interface PoolDetail<
+  TWeight extends string | Big = string,
+  TLiquidity extends Coin | TokenWithValue = Coin
+> extends Pool<TLiquidity> {
+  isSupported: boolean;
+  blockHeight: Option<number>;
+  creator: Option<Addr>;
+  poolAddress: ContractAddr;
+  swapFee: number;
+  exitFee: number;
+  futurePoolGovernor: string | null;
+  smoothWeightChangeParams: object | null;
+  scalingFactors: string[] | null;
+  scalingFactorController: string | null;
+  weight: PoolWeight<TWeight>[] | null;
 }